% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{ae.kd_cv}
\alias{ae.kd_cv}
\title{Absolute Error of kd_cv}
\usage{
\method{ae}{kd_cv}(actual, predicted = join_cv_predictions(actual)$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Absolute Error of kd_cv
}
