% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape.R
\name{ape.default}
\alias{ape.default}
\title{Absolute Percentage Error}
\usage{
\method{ape}{default}(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Absolute Percentage Error
}
