% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape.R
\name{ape.kd_cv}
\alias{ape.kd_cv}
\title{Absolute Percentage Error of kd_cv}
\usage{
\method{ape}{kd_cv}(actual, predicted = actual$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Absolute Percentage Error of kd_cv
}
