% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape.R
\name{ape.kd_model}
\alias{ape.kd_model}
\title{Absolute Percentage Error of kd_model}
\usage{
\method{ape}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Absolute Percentage Error of kd_model
}
