% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias.kd_cv}
\alias{bias.kd_cv}
\title{Bias of kd_cv}
\usage{
\method{bias}{kd_cv}(actual, predicted = join_cv_predictions(actual)$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Bias of kd_cv
}
