% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias.kd_model}
\alias{bias.kd_model}
\title{Bias of kd_model}
\usage{
\method{bias}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Bias of kd_model
}
