% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_model.R
\name{clone_model.default}
\alias{clone_model.default}
\title{Default clone a model instance.}
\usage{
\method{clone_model}{default}(model, input_tensors = NULL, clone_function = NULL, ...)
}
\arguments{
\item{model}{model}

\item{input_tensors}{input_tensors}

\item{clone_function}{clone_function}
}
\description{
Default clone a model instance.
}
