% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_model.R
\name{clone_model.kd_model}
\alias{clone_model.kd_model}
\title{Clone a kd_model instance}
\usage{
\method{clone_model}{kd_model}(model, ...)
}
\arguments{
\item{model}{a kd_model object}

\item{...}{additional parameters}
}
\description{
Clone a kd_model instance
}
