% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_model.R
\name{clone_model}
\alias{clone_model}
\title{Generic clone a model instance.}
\usage{
clone_model(model, input_tensors = NULL, clone_function = NULL, ...)
}
\arguments{
\item{model}{model}

\item{input_tensors}{input_tensors}

\item{clone_function}{clone_function}

\item{...}{additional parameters}
}
\description{
Generic clone a model instance.
}
