% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_response.R
\name{compare_response}
\alias{compare_response}
\title{Compare the actual response to the predicted}
\usage{
compare_response(model, data = model$data)
}
\arguments{
\item{model}{a kd_model object}

\item{data}{a kd_data object, defaults to the data from model}
}
\description{
Compare the actual response to the predicted
}
