% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.plot.r
\name{cor.plot.kd_cv}
\alias{cor.plot.kd_cv}
\title{Correlation Plot for kd_cv}
\usage{
\method{cor.plot}{kd_cv}(x, identity_line = TRUE, ...)
}
\arguments{
\item{x}{a kd_cv object}

\item{identity_line}{include a line of x=y, default TRUE}

\item{...}{additional parameters}
}
\value{
a plot
}
\description{
Correlation Plot for kd_cv
}
