% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.plot.r
\name{cor.plot.kd_model}
\alias{cor.plot.kd_model}
\title{Correlation Plot for kd_model}
\usage{
\method{cor.plot}{kd_model}(x, identity_line = TRUE, ...)
}
\arguments{
\item{x}{a kd_model object}

\item{identity_line}{include a line of x=y, default TRUE}

\item{...}{additional parameters}
}
\value{
a plot
}
\description{
Correlation Plot for kd_model
}
