% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.plot.r
\name{cor.plot.kd_model_htest}
\alias{cor.plot.kd_model_htest}
\title{Correlation Plot for kd_model_htest}
\usage{
\method{cor.plot}{kd_model_htest}(x, ...)
}
\arguments{
\item{x}{a kd_model_htest object}

\item{...}{additional parameters}
}
\value{
a plot
}
\description{
Correlation Plot for kd_model_htest
}
