% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.test.r
\name{cor.test.kd_cv}
\alias{cor.test.kd_cv}
\title{Test for Association/Correlation Between Predicted Rates and Actual Rates in Kedis Cross-Validation}
\usage{
\method{cor.test}{kd_cv}(x, ...)
}
\arguments{
\item{x}{a kd_cv object}

\item{...}{additional parameters}
}
\value{
a htest object
}
\description{
Test for Association/Correlation Between Predicted Rates and Actual Rates in Kedis Cross-Validation
}
