% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.test.r
\name{cor.test.kd_model}
\alias{cor.test.kd_model}
\title{Test for Association/Correlation Between Predicted Rates and Actual Rates in Kedis}
\usage{
\method{cor.test}{kd_model}(x, ...)
}
\arguments{
\item{x}{a kd_model}

\item{...}{additional parameters}
}
\value{
a htest object
}
\description{
Test for Association/Correlation Between Predicted Rates and Actual Rates in Kedis
}
