% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.kd_model}
\alias{cv.kd_model}
\title{Cross-Validation of a kd_model object}
\usage{
\method{cv}{kd_model}(
  object,
  data = object$data,
  k = NA,
  idx = NULL,
  seed = NULL,
  loss = "poisson",
  silent = FALSE,
  ...
)
}
\arguments{
\item{object}{a kd_model object}

\item{data}{data. default data from object. must be a kd_data object.}

\item{k}{k, must include if idx is NULL}

\item{idx}{a kd_cv_index object, must include if k is NA}

\item{seed}{seed}

\item{loss}{the loss function, passed to kedis::loss}

\item{silent}{omit all output}

\item{...}{additional parameters passed to kedis::train}
}
\value{
a kd_cv object
}
\description{
Cross-Validation of a kd_model object
}
