% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.kd_model}
\alias{fit.kd_model}
\title{Train a Kedis model}
\usage{
\method{fit}{kd_model}(
  object,
  x = object$inputs,
  y = object$outputs,
  batch_size = length(y),
  epochs = 1000,
  verbose = getOption("keras.fit_verbose", default = "auto"),
  callbacks = NULL,
  view_metrics = getOption("keras.view_metrics", default = "auto"),
  validation_split = 0,
  validation_data = NULL,
  shuffle = TRUE,
  initial_epoch = 0,
  steps_per_epoch = NULL,
  validation_steps = NULL,
  predictions = FALSE,
  ...
)
}
\arguments{
\item{object}{Model to train. A kd_model object.}

\item{x}{Passed to fit.keras.engine.training.Model. A vector, matrix or array of training data.}

\item{y}{Passed to fit.keras.engine.training.Model. A vector, amtrix or array of target (label) data.}

\item{batch_size}{Passed to fit.keras.engine.training.Model. Integer or NULL. Number of samples per gradient update. If unspecified, batch_size will default to 32.}

\item{epochs}{Number of epochs to train the model. Note that in conjunction with initial_epoch, epochs is to be understood as "final epoch". The model is not trained for a number of iterations given by epochs, but merely until the epoch of index epochs is reached.}

\item{verbose}{Verbosity mode (0 = silent, 1 = progress bar, 2 = one line per epoch).}

\item{callbacks}{List of callbacks to be called during training.}

\item{view_metrics}{View realtime plot of training metrics (by epoch). The default ("auto") will display the plot when running within RStudio, metrics were specified during model compile(), epochs > 1 and verbose > 0. Use the global keras.view_metrics option to establish a different default.}

\item{validation_split}{Float between 0 and 1. Fraction of the training data to be used as validation data. The model will set apart this fraction of the training data, will not train on it, and will evaluate the loss and any model metrics on this data at the end of each epoch. The validation data is selected from the last samples in the x and y data provided, before shuffling.}

\item{validation_data}{Data on which to evaluate the loss and any model metrics at the end of each epoch. The model will not be trained on this data. This could be a list (x_val, y_val) or a list (x_val, y_val, val_sample_weights). validation_data will override validation_split.}

\item{shuffle}{Logical (whether to shuffle the training data before each epoch) or string (for "batch"). "batch" is a special option for dealing with the limitations of HDF5 data; it shuffles in batch-sized chunks. Has no effect when steps_per_epoch is not NULL.}

\item{initial_epoch}{Integer, Epoch at which to start training (useful for resuming a previous training run).}

\item{steps_per_epoch}{Total number of steps (batches of samples) before declaring one epoch finished and starting the next epoch. When training with input tensors such as TensorFlow data tensors, the default NULL is equal to the number of samples in your dataset divided by the batch size, or 1 if that cannot be determined.}

\item{validation_steps}{Only relevant if steps_per_epoch is specified. Total number of steps (batches of samples) to validate before stopping.}

\item{predictions}{Boolean. Should predictions be made at each epoch. Useful to visualise training. Very slow. Defaults to FALSE.}

\item{...}{Passed to fit.keras.engine.training.Model. Unused.}
}
\value{
A kd_history object that contains all information collected during training.
}
\description{
Train a Kedis model
}
