% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subset.r
\name{get_subset.kd_data}
\alias{get_subset.kd_data}
\title{Get subset of kd_data}
\usage{
\method{get_subset}{kd_data}(data, idx, n_out, n_in = NULL, ...)
}
\arguments{
\item{data}{a kd_data object}

\item{idx}{a kd_index object}

\item{n_out}{number of outer loops}

\item{n_in}{number of inner loops (NULL for non-nested Cross-Validation)}

\item{...}{additional parameters}
}
\value{
either a kd_cv_data or a kd_ncv_data, depending on the index and n_in
}
\description{
Get subset of kd_data
}
