% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subset.r
\name{get_subset.kd_ncv_index}
\alias{get_subset.kd_ncv_index}
\title{Get subset of kd_data with kd_ncv_index}
\usage{
\method{get_subset}{kd_ncv_index}(data, idx, n_out, n_in = NULL, ...)
}
\arguments{
\item{data}{a kd_data object}

\item{idx}{a kd_ncv_index object}

\item{n_out}{outer loop}

\item{n_in}{inner loop}

\item{...}{additional parameters}
}
\value{
a kd_ncv_data object, which is a list of kd_data objects
}
\description{
Get subset of kd_data with kd_ncv_index
}
