% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypers_str.r
\name{hypers_str}
\alias{hypers_str}
\title{Kedis hyperparameters to string}
\usage{
hypers_str(hypers)
}
\arguments{
\item{hypers}{lsit of keras layers}
}
\value{
a string
}
\description{
Kedis hyperparameters to string
}
