% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_cv_predictions.R
\name{join_cv_predictions}
\alias{join_cv_predictions}
\title{Joins Kedis Cross-Validation predicted rates with actual rates}
\usage{
join_cv_predictions(x, ...)
}
\arguments{
\item{x}{a kd_cv object}

\item{...}{additional parameters}
}
\value{
a data.frame
}
\description{
Joins Kedis Cross-Validation predicted rates with actual rates
}
