% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss.kd_model}
\alias{loss.kd_model}
\title{loss}
\usage{
\method{loss}{kd_model}(model, data = model$data, full = FALSE, min = TRUE, ...)
}
\arguments{
\item{model}{a kd_model object}

\item{data}{a kd_data object, defaults to data from model}

\item{full}{return the full dataframe, default FALSE}

\item{min}{include the minimum (optimum) loss}

\item{...}{additional parameters}
}
\value{
a data.frame
}
\description{
loss
}
