% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae.r
\name{mae.kd_cv}
\alias{mae.kd_cv}
\title{Mean Absolute Error of kd_cv}
\usage{
\method{mae}{kd_cv}(actual, predicted = actual$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Mean Absolute Error of kd_cv
}
