% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae.r
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Mean Absolute Error
}
