% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{mape.kd_cv}
\alias{mape.kd_cv}
\title{Mean Absolute Percentage Error of kd_cv}
\usage{
\method{mape}{kd_cv}(actual, predicted = actual$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Mean Absolute Percentage Error of kd_cv
}
