% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{mape.kd_model}
\alias{mape.kd_model}
\title{Mean Absolute Percentage Error of kd_model}
\usage{
\method{mape}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Mean Absolute Percentage Error of kd_model
}
