% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdae.R
\name{mdae.default}
\alias{mdae.default}
\title{Median Absolute Error}
\usage{
\method{mdae}{default}(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Median Absolute Error
}
