% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdae.R
\name{mdae.kd_cv}
\alias{mdae.kd_cv}
\title{Median Absolute Error of kd_cv}
\usage{
\method{mdae}{kd_cv}(actual, predicted = join_cv_predictions(actual)$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Median Absolute Error of kd_cv
}
