% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse.kd_cv}
\alias{mse.kd_cv}
\title{Mean Squared Error of kd_cv}
\usage{
\method{mse}{kd_cv}(actual, predicted = actual$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Mean Squared Error of kd_cv
}
