% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse.kd_model}
\alias{mse.kd_model}
\title{Mean Squared Error of kd_model}
\usage{
\method{mse}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Mean Squared Error of kd_model
}
