% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msle.R
\name{msle}
\alias{msle}
\title{Mean Squared Log Error}
\usage{
msle(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Mean Squared Log Error
}
