% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncv.R
\name{ncv.kd_data}
\alias{ncv.kd_data}
\title{Nested Cross Validation for Kedis}
\usage{
\method{ncv}{kd_data}(
  data,
  n_out_loop,
  n_in_loop,
  hypers,
  seed,
  csv_folder = NA,
  out_loop = NA,
  ...
)
}
\arguments{
\item{data}{a kd_data object}

\item{n_out_loop}{number of outer loops}

\item{n_in_loop}{number of inner loops}

\item{hypers}{list of hyperparameters}

\item{seed}{seed}

\item{csv_folder}{folder address to save results csv to. if na, will not save, but return a data frame}

\item{out_loop}{which outer loop to run, defaults to all}

\item{...}{additional parameters}
}
\value{
a data.frame and a kd_ncv_index object
}
\description{
Nested Cross Validation for Kedis
}
