% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncv.R
\name{ncv.kd_model}
\alias{ncv.kd_model}
\title{Nested Cross-Validation for a Kedis Model}
\usage{
\method{ncv}{kd_model}(
  model,
  data = model$data,
  n_out_loop,
  n_in_loop,
  hypers,
  loss = "poisson",
  seed = NULL,
  file_path = NULL,
  silent = FALSE,
  ...
)
}
\arguments{
\item{model}{a kd_model to be used as template}

\item{data}{kd_data, if not supplied data from kd_model is used}

\item{n_out_loop}{number of outer loops}

\item{n_in_loop}{numer of inner loops}

\item{hypers}{list of hyperparameter sets}

\item{loss}{loss to calculate, passed to kedis::loss}

\item{seed}{random seed}

\item{file_path}{file path to save csv file. if NULL file will not be saved}

\item{silent}{omit all output}

\item{...}{additional parameters to pass to train}
}
\description{
Nested Cross-Validation for a Kedis Model
}
