% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_for_predict.R
\name{new_data_for_predict}
\alias{new_data_for_predict}
\title{Format new data to predict using a kedis model}
\usage{
new_data_for_predict(model, covariates, population)
}
\arguments{
\item{model}{the kd_model object for prediction (will not work if xy is included as covariates)}

\item{covariates}{a SpatRaster of covariates (must be the same covariates as in the model)}

\item{population}{a SpatRaster of population}
}
\description{
Format new data to predict using a kedis model
}
