% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_zeros.r
\name{pad_zeros}
\alias{pad_zeros}
\title{Append rows of zeros to a data.frame}
\usage{
pad_zeros(df, length, pad = 0)
}
\arguments{
\item{df}{the dataframe}

\item{length}{total length after padding}

\item{pad}{value to pad with, default 0}
}
\description{
Append rows of zeros to a data.frame
}
