% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_bias.R
\name{percent_bias.default}
\alias{percent_bias.default}
\title{Percent Bias}
\usage{
\method{percent_bias}{default}(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Percent Bias
}
