% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_bias.R
\name{percent_bias.kd_cv}
\alias{percent_bias.kd_cv}
\title{Percent Bias of kd_cv}
\usage{
\method{percent_bias}{kd_cv}(actual, predicted = join_cv_predictions(actual)$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Percent Bias of kd_cv
}
