% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_bias.R
\name{percent_bias.kd_model}
\alias{percent_bias.kd_model}
\title{Percent Bias of kd_model}
\usage{
\method{percent_bias}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Percent Bias of kd_model
}
