% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.kd_cv}
\alias{plot.kd_cv}
\title{Plot the history of a kd_cv object}
\usage{
\method{plot}{kd_cv}(x, ...)
}
\arguments{
\item{x}{a kd_cv object}

\item{...}{parameters to pass to plot.keras_training_history}
}
\description{
Plot the history of a kd_cv object
}
