% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.kd_model}
\alias{plot.kd_model}
\title{plot}
\usage{
\method{plot}{kd_model}(
  x,
  data = x$data,
  gglayers = NULL,
  reagg = FALSE,
  xy = FALSE,
  compare = FALSE,
  ...
)
}
\arguments{
\item{x}{kd_model}

\item{data}{kd_data, defaults to data from kd_model}

\item{gglayers}{list of layers to add to ggplot}

\item{reagg}{reaggregate output}

\item{xy}{plot field, only valid if layers_xy included otherwise returns prediction}

\item{compare}{compare to actual data, only works tfor reagg = TRUE}

\item{...}{for plot generic}
}
\value{
a kd_plot object
}
\description{
plot
}
