% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.kd_new_data}
\alias{plot.kd_new_data}
\title{Predict then plot a kedis model with new data}
\usage{
\method{plot}{kd_new_data}(x, ...)
}
\arguments{
\item{x}{a kd_new_data object, from new_data_for_predict function}

\item{...}{additional parameters passed to predict}
}
\description{
Predict then plot a kedis model with new data
}
