% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.kd_new_data_predict_df}
\alias{plot.kd_new_data_predict_df}
\title{Plot a kedis model with new data}
\usage{
\method{plot}{kd_new_data_predict_df}(x, ...)
}
\arguments{
\item{x}{a kd_new_data_predict_df object}

\item{...}{additional parameters}
}
\description{
Plot a kedis model with new data
}
