% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.kd_predict}
\alias{plot.kd_predict}
\title{plot a kd_predict}
\usage{
\method{plot}{kd_predict}(x, layer = "output_disag", ...)
}
\arguments{
\item{x}{a kd_predict object, made from predict.kd_model}

\item{layer}{which layer to plot. One of either "output_disag", "output_agg", "output_xy", "output_rate", or any column from the "shapes" SpatVector}

\item{...}{for plot generic}
}
\description{
plot a kd_predict
}
