% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict.kd_model}
\alias{predict.kd_model}
\title{predict}
\usage{
\method{predict}{kd_model}(object, data = object$data, as.data.frame = FALSE, ...)
}
\arguments{
\item{object}{a kd_model object}

\item{data}{a kd_data object. defaults to the data from object}

\item{as.data.frame}{If FALSE, returns as SpatRasters or SpatVectors. If TRUE, returns as data.frames for rasters and sf for vectors}

\item{...}{for predict generic}
}
\value{
a kd_predict object
}
\description{
predict
}
