% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict.kd_new_data}
\alias{predict.kd_new_data}
\title{Predict a kedis model with new data}
\usage{
\method{predict}{kd_new_data}(object, as.data.frame = FALSE, ...)
}
\arguments{
\item{object}{a kd_new_data object, created by the new_data_for_predict function in kedis}

\item{as.data.frame}{If FALSE, returns as SpatRasters. If TRUE, returns as data.frame}

\item{...}{additional parameters for predict}
}
\description{
Predict a kedis model with new data
}
