% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.kd_model}
\alias{print.kd_model}
\title{Print kd_model}
\usage{
\method{print}{kd_model}(x, ...)
}
\arguments{
\item{x}{a kd_model object}

\item{...}{additional parameters}
}
\description{
Print kd_model
}
