% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/summary.R
\name{print.kd_ncv}
\alias{print.kd_ncv}
\title{Print kd_ncv}
\usage{
\method{print}{kd_ncv}(object, ...)

\method{print}{kd_ncv}(object, ...)
}
\arguments{
\item{object}{a kd_ncv object}

\item{...}{additional parameters}

\item{x}{a kd_ncv object}
}
\description{
Print kd_ncv

Summarise kd_ncv
}
