% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rae.R
\name{rae.kd_cv}
\alias{rae.kd_cv}
\title{Relative Absolute Error of kd_cv}
\usage{
\method{rae}{kd_cv}(actual, predicted = actual$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Relative Absolute Error of kd_cv
}
