% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rae.R
\name{rae}
\alias{rae}
\title{Relative Absolute Error}
\usage{
rae(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Relative Absolute Error
}
