% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse.kd_model}
\alias{rmse.kd_model}
\title{Root Mean Squared Error of kd_model}
\usage{
\method{rmse}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Root Mean Squared Error of kd_model
}
