% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsle.R
\name{rmsle.default}
\alias{rmsle.default}
\title{Root Mean Squared Log Error}
\usage{
\method{rmsle}{default}(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Root Mean Squared Log Error
}
