% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrse.R
\name{rrse.kd_cv}
\alias{rrse.kd_cv}
\title{Root Relative Squared Error of kd_cv}
\usage{
\method{rrse}{kd_cv}(actual, predicted = actual$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Root Relative Squared Error of kd_cv
}
