% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrse.R
\name{rrse.kd_model}
\alias{rrse.kd_model}
\title{Root Relative Squared Error of kd_model}
\usage{
\method{rrse}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Root Relative Squared Error of kd_model
}
