% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrse.R
\name{rrse}
\alias{rrse}
\title{Root Relative Squared Error}
\usage{
rrse(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Root Relative Squared Error
}
