% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse.R
\name{rse.kd_model}
\alias{rse.kd_model}
\title{Relative Squared Error of kd_model}
\usage{
\method{rse}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Relative Squared Error of kd_model
}
