% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse.R
\name{rse}
\alias{rse}
\title{Relative Squared Error}
\usage{
rse(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Relative Squared Error
}
