% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se.kd_cv}
\alias{se.kd_cv}
\title{Squared Error of kd_cv}
\usage{
\method{se}{kd_cv}(actual, predicted = actual$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Squared Error of kd_cv
}
