% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se.kd_model}
\alias{se.kd_model}
\title{Squared Error of kd_model}
\usage{
\method{se}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Squared Error of kd_model
}
