% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sle.R
\name{sle.default}
\alias{sle.default}
\title{Squared Log Error}
\usage{
\method{sle}{default}(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Squared Log Error
}
