% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sle.R
\name{sle.kd_model}
\alias{sle.kd_model}
\title{Squared Log Error of kd_model}
\usage{
\method{sle}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Squared Log Error of kd_model
}
