% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smape.R
\name{smape.default}
\alias{smape.default}
\title{Symmetric Mean Absolute Percentage Error}
\usage{
\method{smape}{default}(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector is a prediction for the corresponding element in actual.}
}
\description{
Symmetric Mean Absolute Percentage Error
}
