% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smape.R
\name{smape.kd_cv}
\alias{smape.kd_cv}
\title{Symmetric Mean Absolute Percentage Error of kd_cv}
\usage{
\method{smape}{kd_cv}(actual, predicted = join_cv_predictions(actual)$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Symmetric Mean Absolute Percentage Error of kd_cv
}
