% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smape.R
\name{smape.kd_model}
\alias{smape.kd_model}
\title{Symmetric Mean Absolute Percentage Error of kd_model}
\usage{
\method{smape}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Symmetric Mean Absolute Percentage Error of kd_model
}
