% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse.R
\name{sse.kd_cv}
\alias{sse.kd_cv}
\title{Sum of Squared Errors of kd_cv}
\usage{
\method{sse}{kd_cv}(actual, predicted = actual$predicted)
}
\arguments{
\item{actual}{a fitted kd_cv object}

\item{predicted}{the predicted rates, default is from kd_cv}
}
\description{
Sum of Squared Errors of kd_cv
}
