% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse.R
\name{sse.kd_model}
\alias{sse.kd_model}
\title{Sum of Squared Errors of kd_model}
\usage{
\method{sse}{kd_model}(
  actual,
  predicted = predict(actual$predict_model, actual$data$inputs, verbose = 0)[[4]]
)
}
\arguments{
\item{actual}{a fitted kd_model object}

\item{predicted}{the predicted rates, default is from kd_model}
}
\description{
Sum of Squared Errors of kd_model
}
