% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.kd_cv}
\alias{summary.kd_cv}
\title{Summarize kd_cv}
\usage{
\method{summary}{kd_cv}(object, ...)
}
\arguments{
\item{object}{a kd_cv object}

\item{...}{additional parameters}
}
\description{
Summarize kd_cv
}
