% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.kd_data}
\alias{summary.kd_data}
\title{Summarize kd_data}
\usage{
\method{summary}{kd_data}(object, ...)
}
\arguments{
\item{object}{a kd_data object}

\item{...}{additional parameters}
}
\description{
Summarize kd_data
}
