% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.kd_model}
\alias{summary.kd_model}
\title{Summarize kd_model}
\usage{
\method{summary}{kd_model}(object, ...)
}
\arguments{
\item{object}{a kd_model object}

\item{...}{additional parameters}
}
\description{
Summarize kd_model
}
