% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.kd_ncv}
\alias{summary.kd_ncv}
\title{Summarise kd_ncv}
\usage{
\method{summary}{kd_ncv}(object, ...)
}
\arguments{
\item{object}{a kd_ncv object}

\item{...}{additional parameters}
}
\description{
Summarise kd_ncv
}
