% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss.kd_model}
\alias{loss.kd_model}
\title{loss}
\usage{
\method{loss}{kd_model}(model, data = model$data, loss = "poisson", ...)
}
\arguments{
\item{model}{a kd_model object}

\item{data}{a kd_data object, defaults to data from model}

\item{loss}{the loss function to apply. either a string "poisson" (default) which is the poisson loss, or any function which takes actual and predicted as parameters}

\item{...}{additional parameters}
}
\value{
a data.frame
}
\description{
loss
}
