% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncv.R
\name{ncv}
\alias{ncv}
\title{Nested cross Validation}
\usage{
ncv(model, ...)
}
\arguments{
\item{model}{the model}

\item{...}{additional parameters}
}
\value{
the results, duh...
}
\description{
Nested cross Validation
}
