% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.kd_model}
\alias{plot.kd_model}
\title{Plot a kd-model}
\usage{
\method{plot}{kd_model}(x, layers = "output_disag", data = x$data, ...)
}
\arguments{
\item{x}{a kd_model object}

\item{layers}{which layer to plot. One of either "output_disag", "output_agg", "output_xy", "output_rate", or any column from the "shapes" SpatVector}

\item{data}{a kd_Data object, default to data from x}

\item{...}{additional parameters to pass to predict}
}
\description{
Plot a kd-model
}
