% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.kd_model}
\alias{predict.kd_model}
\title{Generate predictions from a Kedis model}
\usage{
\method{predict}{kd_model}(
  object,
  batch_size = NULL,
  verbose = "auto",
  steps = NULL,
  callbacks = NULL,
  as_terra = TRUE,
  ...
)
}
\arguments{
\item{object}{Kedis model.}

\item{batch_size}{Integer. If unspecified, it will default to 32.}

\item{verbose}{Verbosity mode, 0, 1, 2, or "auto". "auto" defaults to 1 for for most cases and defaults to verbose=2 when used with ParameterServerStrategy or with interactive logging disabled.}

\item{steps}{Total number of steps (batches of samples) before declaring the evaluation round finished. Ignored with the default value of NULL.}

\item{callbacks}{List of callbacks to apply during prediction.}

\item{as_terra}{Convert prediciton to SpatRaster and SpatVector. Default TRUE.}

\item{...}{Passed to predict.keras.engine.training.Model. Unused.}
}
\value{
A kd_prediction object.
}
\description{
Generate predictions from a Kedis model
}
