% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict.kd_model}
\alias{predict.kd_model}
\title{predict}
\usage{
\method{predict}{kd_model}(object, data = object$data, reagg = FALSE, xy = FALSE, ...)
}
\arguments{
\item{object}{kd_model}

\item{data}{kd_data}

\item{reagg}{reaggregate output}

\item{xy}{return field output, only works if model has layers_xy declared, otherwise returns prediction}

\item{...}{for predict generic}
}
\value{
a kd_predict object
}
\description{
predict
}
