% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(
  shapes,
  covariates,
  population = NULL,
  filter_var,
  response_var,
  cov_names = NULL,
  na.action = TRUE,
  cov_missing = function(.x) {
     median(.x, na.rm = TRUE)
 },
  pop_missing = function(.x) {
     0
 }
)
}
\arguments{
\item{shapes}{SpatVector of shapefile, or address to .shp file}

\item{covariates}{SpatRaster of covariates, or address of .tif or folder of .tifs.}

\item{population}{SpatRaster of population, or addtess of .tif}

\item{filter_var}{Variable in shape file to filter by}

\item{response_var}{Response in shape file}

\item{cov_names}{Covariates to include, defaults to all covariates in covariates argument. Can include x and y}

\item{na.action}{If true, fills missing data on rasters (using formula cov_missing and pop_missing). If false, only pixels with non-missing values on all layers are included. Default TRUE.}

\item{cov_missing}{formula for missing covariates. performed column-wise}

\item{pop_missing}{formulat for missing population. performed column-wise}
}
\value{
a kd_data object
}
\description{
prepare_data
}
