% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{train.kd_model}
\alias{train.kd_model}
\title{train}
\usage{
\method{train}{kd_model}(x, data = x$data, validation_data = NULL, ..., verbose = 0)
}
\arguments{
\item{x}{kd_model}

\item{data}{kd_data}

\item{validation_data}{validation data as a kd_data object}

\item{...}{arguments to keras::fit}

\item{verbose}{verbosity, defaults to 0}
}
\value{
a kd_history object
}
\description{
train
}
